# testplanit/lib/openapi/custom-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TestPlanIt Custom API",
    "version": "1.0.0",
    "description": "Custom API endpoints for TestPlanIt (non-ZenStack endpoints)"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Server"
    }
  ],
...
  "paths": {
...
    "/projects/{projectId}/cases/bulk-create": {
      "post": {
        "tags": ["Projects", "Repository"],
        "summary": "Bulk create test cases",
        "description": "Creates up to 200 test cases in one call, grouped by their effective folder and workflow state so each group lands in a single transaction. Each case gets its steps, tags, custom field values and a version 1 snapshot.\n\nPartial failure is reported per case rather than failing the batch: `results` carries one entry per submitted case, keyed by its index in `cases`. A transaction failure fails every case in that folder/state group. Accepts a session cookie or a Bearer API token; a token scoped `mode:read` is rejected with 403.",
        "operationId": "bulkCreateCases",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["folderId", "cases"],
                "properties": {
                  "folderId": {
                    "type": "integer",
                    "description": "Default folder for the batch. Must belong to this project. Each case may override it."
                  },
                  "templateId": {
                    "type": "integer",
                    "description": "Template for the batch. Defaults to the project's first enabled assigned template."
                  },
                  "stateName": {
                    "type": "string",
                    "description": "Default CASES-scope workflow state name. Defaults to the first state by order. Each case may override it."
                  },
                  "cases": {
                    "type": "array",
                    "minItems": 1,
                    "maxItems": 200,
                    "items": {
                      "type": "object",
                      "required": ["name"],
                      "properties": {
                        "name": {
                          "type": "string",
                          "minLength": 1,
                          "maxLength": 2000
                        },
                        "folderId": { "type": "integer" },
                        "stateName": { "type": "string" },
                        "steps": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "text": { "type": "string" },
                              "expectedResult": { "type": "string" },
                              "order": { "type": "integer", "minimum": 0 }
                            }
                          }
                        },
                        "tags": {
                          "type": "array",
                          "description": "Tag IDs, or tag names which are created if missing.",
                          "items": {
                            "oneOf": [
                              { "type": "integer" },
                              { "type": "string" }
                            ]
                          }
                        },
                        "customFields": {
                          "type": "object",
                          "description": "Custom field values keyed by the template field's display name. A name not on the template fails that case.",
                          "additionalProperties": true
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch processed. Inspect `results` for per-case outcomes; `failedCount` may be non-zero.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "importedCount": { "type": "integer" },
                    "failedCount": { "type": "integer" },
                    "results": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "description": "The case's index in the submitted `cases` array."
                          },
                          "name": { "type": "string" },
                          "status": {
                            "type": "string",
                            "enum": ["success", "error"]
                          },
                          "caseId": {
                            "type": "integer",
                            "description": "Present on success."
                          },
                          "error": {
                            "type": "string",
                            "description": "Present on failure."
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid project ID, invalid body, or a template that is not enabled and assigned to the project"
          },
          "401": { "description": "No session and no valid API token" },
          "403": { "description": "API token is scoped `mode:read`" },
          "404": { "description": "Project not found or access denied" },
          "422": {
            "description": "No enabled template is assigned to the project"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth.js session cookie"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token authentication"
      }
    }
  },
  "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }]
}